fix: 가격 타입 long 통일, DB 스키마 BIGINT 변경 및 DTO 검증 개선

+++ src/main/java/buncheoleasy/buncheol/domain/BuncheolParams.java
@@ -9,7 +9,7 @@ public record BuncheolParams(
         String description,
         String goodsName,
         String storeName,
-        int originalPrice,
+        long originalPrice,
         LocalDateTime deadline,
         int shippingDeadlineDays,
         Integer gs25ShippingFee,
